Refactor get_possible_concrete_types into utils/inspect.py
